Solving duplications
adding additional service to avoid duplications
