+++ src/intentEmitters/intentListeners/kerotoplasty.bot.condition.Identification.listener.ts
@@ -1,8 +1,7 @@
 
+import { sendExtraMessages } from "../../services/send.extra.messages.service";
 import { Logger } from "../../common/logger";
 import { kerotoplastyService } from "../../services/kerotoplasty.service";
-import { getAdditionalInfoSevice } from "../../services/get.additional.info.service";
-import { translateService } from "../../services/translate.service";
 
 export const kerotoplastyConditionIdentificationListener = async (intent, eventObj) => {
 
@@ -10,7 +9,7 @@ export const kerotoplastyConditionIdentificationListener = async (intent, eventO
     try {
         const condition = await kerotoplastyServiceObj.identifyCondition(eventObj);
         const [response,severityGrade] = await kerotoplastyServiceObj.conditionSpecificResponse(condition);
-        keratoplastyNextSteps(intent,eventObj,severityGrade);
+        askingForImage(intent,eventObj,severityGrade);
         return response;
     }
     catch (error) {
@@ -19,25 +18,17 @@ export const kerotoplastyConditionIdentificationListener = async (intent, eventO
         throw new Error("Keratoplasty bot condition identification Listener Error");
     }
 };
-async function keratoplastyNextSteps(intent,eventObj,severityGrade) {
+async function askingForImage(intent,eventObj,severityGrade) {
     try {
-        console.log("STEP 4");
         const kerotoplastyServiceObj: kerotoplastyService = eventObj.container.resolve(kerotoplastyService);
-        let message = "Would you like to send an image of the affected area for the doctor's reference?";
-        const additionalObj: getAdditionalInfoSevice = eventObj.container.resolve(getAdditionalInfoSevice);
-        const userId = eventObj.body.originalDetectIntentRequest.payload.userId;
-        const languageCode = eventObj.body.queryResult.languageCode;
-        const translationServiceObj: translateService = eventObj.container.resolve( translateService);
-        const button_yes = await translationServiceObj.translatestring("Yes",languageCode);
-        const button_no = await translationServiceObj.translatestring("No",languageCode);
-        message  = await translationServiceObj.translatestring(message,languageCode);
-        const buttonArray = [button_yes, "EyeImage",button_no,"responseNo"];
-        additionalObj.sendResponsebyButton( message,eventObj, userId,buttonArray);
+        const sendExtraMessagesobj: sendExtraMessages = eventObj.container.resolve(sendExtraMessages);
+        const inputMessage =  "Would you like to send an image of the affected area for the doctor's reference?";
+        const yesIntentName = "EyeImage";
+        const noIntentName = "responseNo";
+        sendExtraMessagesobj.sendSecondaryButtonMessage(inputMessage, yesIntentName, noIntentName,  eventObj);
         kerotoplastyServiceObj.postingOnClickup(intent,eventObj,severityGrade);
     } catch (error) {
         console.log(error);
         throw new Error("Keratoplasty next steps error");
     }
-
-    
 }
